Python Modules/CPSL_TI_Radar/pyproject.toml
[tool.poetry]
name = "cpsl-ti-radar"
version = "0.1.0"
description = "Python Packaged to integrate CPSL hardware with TI radars and data capture cards"
authors = ["David <david.hunt@duke.edu>"]
license = "MIT"
readme = "README.md"
packages = [{include = "CPSL_TI_Radar"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pyserial = "^3.5"
matplotlib = "^3.7.2"
scipy = "^1.11.1"
numpy = "^1.25.1"
imageio = {extras = ["ffmpeg"], version = "^2.31.1"}
pyqt5 = "^5.15.9"
jupyterlab = "^4.0.3"
pandas = "^2.0.3"
ipykernel = "^6.25.0"
spyder = "^5.4.4"
scikit-learn = "^1.3.0"
tqdm = "^4.65.0"
natsort = "^8.4.0"
open3d = "^0.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
